requirements.txt
django==3.2
djangorestframework
pyTelegramBotAPI
gunicorn
drf-yasg
psycopg2-binary
django-cors-headers